LayoutTests/media/video-autoplay-allowed-but-fullscreen-required-expected.txt

RUN(internals.settings.setAllowsInlineMediaPlayback(false))
EVENT(canplaythrough)
Video did not begin playing OK
END OF TEST
